+++ ts-force-gen/src/util.ts
@@ -3,12 +3,12 @@ import * as fs from 'fs';
 
 const API_NAME_REGEX = /(?:^((?:\w(?!__))+\w)__|^)((?:\w(?!__))+\w)(?:__(.+)$|$)/;
 
-export const cleanAPIName = (sfName: string, stripNamespace: boolean) => {
+export const cleanAPIName = (sfName: string, keepNamespaces: boolean) => {
   let match = API_NAME_REGEX.exec(sfName);
   if (!match) {
     throw new Error('NO MATCH FOUND FOR ' + sfName);
   }
-  let name = (!stripNamespace && match[1] ? match[1] : '') + match[2];
+  let name = (keepNamespaces && match[1] ? match[1] : '') + match[2];
   const parts = name.split('_');
   return parts.map((p, i) => {
     if(i > 0 && p.length) {