+++ ts-force-gen/test/generatedSobsTests.spec.ts
@@ -7,10 +7,11 @@ import { cleanAPIName } from '../src/util';
 @suite class TestGenerator {
 
   @test async 'Sanitize Names' () {
-    expect(cleanAPIName('My_Test_Object__c', true)).to.equal('MyTestObject');
-    expect(cleanAPIName('My_Test_Relation__r', false)).to.equal('MyTestRelation');
-    expect(cleanAPIName('NS__Test_Object__r', true)).to.equal('TestObject');
-    expect(cleanAPIName('NS__Test_Object__r', false)).to.equal('NSTestObject');
-    expect(cleanAPIName('NS__Test_object__r', false)).to.equal('NSTestObject');
+    expect(cleanAPIName('My_Test_Object__c', false)).to.equal('MyTestObject');
+    expect(cleanAPIName('My_Test_Relation__r', true)).to.equal('MyTestRelation');
+    expect(cleanAPIName('NS__Test_Object__r', false)).to.equal('TestObject');
+    expect(cleanAPIName('NS__Test_Object__r', true)).to.equal('NSTestObject');
+    expect(cleanAPIName('NS__Test_object__r', true)).to.equal('NSTestObject');
+    expect(cleanAPIName('NS__Test_object__r', true)).to.equal('NSTestObject');
   }
 }